+++ utility.cpp
@@ -1,4 +1,4 @@
-inline int clampData(int min, int data, int max) {
+inline float clampData(float min, float data, float max) {
     if(data < min)return min;
     if(data > max)return max;
     return data;
